# server/openapi.yaml
openapi: 3.1.0
info:
  title: QFree API
  version: 1.0.0
  description: |
    Smart healthcare queue management and real-time waiting-time information.

    **Auth:** send `Authorization: Bearer <accessToken>` (15 min). Renew with `POST /auth/refresh`,
    which reads the httpOnly `qf_rt` cookie and rotates it.

    **Errors:** `{ "error": { "code": "QUEUE_FULL", "message": "…", "details": … } }`.

    **Real-time:** Socket.IO on the same origin. Handshake `auth: { token }` (optional).
    Client → server: `queue:watch`, `queue:unwatch`, `queue:watch-staff`, `queue:unwatch-staff`, `org:watch` (arg: id, ack `{ok, error?}`).
    Server → client: `queue:snapshot` (QueueSnapshot), `queue:entry` (EntryView), `queue:staff` (StaffView), `org:queue`, `notification:new`.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: qf_rt }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    entryId: { name: entryId, in: path, required: true, schema: { type: string, format: uuid } }
    page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    pageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
    from: { name: from, in: query, schema: { type: string, format: date }, description: Default 29 days ago }
    to: { name: to, in: query, schema: { type: string, format: date }, description: Default today }
    bucket: { name: bucket, in: query, schema: { type: string, enum: [day, week, month], default: day } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NoContent: { description: No content }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, QUEUE_FULL, QUEUE_CLOSED, QUEUE_PAUSED, ALREADY_IN_QUEUE, OUTSIDE_HOURS, DOCTOR_UNAVAILABLE, TOO_MANY_ACTIVE_QUEUES, RATE_LIMITED] }
            message: { type: string }
            details: {}
...
    QueueSnapshot:
      type: object
      description: Public, anonymous-safe queue state (no patient identities).
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tokenPrefix: { type: string, example: QF }
        status: { type: string, enum: [OPEN, PAUSED, CLOSED] }
        organization: { type: object }
        doctor: { type: [object, 'null'] }
        service: { type: [object, 'null'] }
        sessionDate: { type: [string, 'null'], example: '2026-09-27' }
        currentToken: { type: [string, 'null'], example: QF-024 }
        waitingCount: { type: integer, example: 6 }
        servedCount: { type: integer }
        capacity: { type: integer }
        remainingCapacity: { type: integer }
        avgServiceMinutes: { type: number }
        estimatedWaitMinutes: { type: integer, description: Wait for someone joining now }
        isAcceptingPatients: { type: boolean }
        updatedAt: { type: string, format: date-time }
    EntryView:
      type: object
      description: The caller's own position in a queue.
      properties:
        queueId: { type: string, format: uuid }
        entry:
          type: object
          properties:
            id: { type: string, format: uuid }
            tokenLabel: { type: string, example: QF-031 }
            status: { type: string, enum: [WAITING, SERVING, COMPLETED, SKIPPED, CANCELLED, NO_SHOW] }
            priority: { type: string, enum: [NORMAL, PRIORITY, EMERGENCY] }
            joinedAt: { type: string, format: date-time }
        currentToken: { type: [string, 'null'], example: QF-024 }
        patientsAhead: { type: integer, example: 6 }
        estimatedWaitMinutes: { type: integer, example: 35 }
        expectedAt: { type: [string, 'null'], format: date-time }
        phase: { type: string, enum: [YOUR_TURN, NEXT, APPROACHING, WAITING, DONE, SKIPPED, CANCELLED, CLOSED] }
        progress: { type: number, minimum: 0, maximum: 1 }
    StaffView:
      type: object
      properties:
        snapshot: { $ref: '#/components/schemas/QueueSnapshot' }
        serving: { type: [object, 'null'] }
        waiting: { type: array, items: { type: object } }
        skipped: { type: array, items: { type: object } }
        finished: { type: array, items: { type: object } }
        stats: { type: object }
    QueueInput:
      type: object
      required: [organizationId, name]
      properties:
        organizationId: { type: string, format: uuid }
        doctorId: { type: [string, 'null'], format: uuid }
        serviceId: { type: [string, 'null'], format: uuid }
        name: { type: string }
        tokenPrefix: { type: string, pattern: '^[A-Z]{1,4}$', default: QF }
        capacity: { type: integer, minimum: 1, maximum: 5000, default: 100 }
        avgServiceMinutes: { type: number, minimum: 1, maximum: 60, default: 10 }
        approachingThreshold: { type: integer, minimum: 1, maximum: 20, default: 3 }
...
security:
  - bearer: []
paths:
...
  /queues:
    get:
      tags: [Queues]
      security: []
      parameters:
        - { name: organizationId, in: query, schema: { type: string, format: uuid } }
        - { name: doctorId, in: query, schema: { type: string, format: uuid } }
        - { name: serviceId, in: query, schema: { type: string, format: uuid } }
      responses: { '200': { description: '{ items: QueueSnapshot[] }' } }
    post:
      tags: [Queue operations]
      summary: Create a queue (org OWNER/MANAGER, or the doctor for their own queue)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/QueueInput' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/QueueSnapshot' } } } }, '400': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
  /queues/{id}:
    parameters: [{ $ref: '#/components/parameters/id' }]
    get: { tags: [Queues], security: [], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QueueSnapshot' } } } }, '404': { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Queue operations], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/QueueInput' } } } }, responses: { '200': { description: OK } } }
    delete: { tags: [Queue operations], summary: Archive (must be closed), responses: { '204': { $ref: '#/components/responses/NoContent' }, '409': { $ref: '#/components/responses/Error' } } }
  /queues/{id}/status:
    get:
      tags: [Queues]
      security: [{}, { bearer: [] }]
      summary: Snapshot plus the caller's own entry when authenticated
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses: { '200': { description: '{ snapshot: QueueSnapshot, myEntry: EntryView | null }' } }
  /queues/{id}/me:
    get: { tags: [Queues], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: '{ myEntry: EntryView | null }' } } }
  /queues/{id}/join:
    post:
      tags: [Queues]
      summary: Join as a patient and receive the next token
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses:
        '201': { description: Joined, content: { application/json: { schema: { $ref: '#/components/schemas/EntryView' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '409': { description: 'QUEUE_CLOSED · QUEUE_FULL · ALREADY_IN_QUEUE · DOCTOR_UNAVAILABLE · TOO_MANY_ACTIVE_QUEUES', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /queues/{id}/leave:
    delete: { tags: [Queues], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: Left }, '409': { $ref: '#/components/responses/Error' } } }
  /queues/{id}/staff:
    get: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/StaffView' } } } }, '403': { $ref: '#/components/responses/Error' } } }
  /queues/{id}/open:
    post:
      tags: [Queue operations]
      summary: Open (or start a new day's session); resets tokens on a new day
      parameters: [{ $ref: '#/components/parameters/id' }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { override: { type: boolean, description: Open outside operating hours (audited) } } } } } }
      responses: { '200': { description: OK }, '409': { description: OUTSIDE_HOURS · DOCTOR_UNAVAILABLE · ORG_INACTIVE } }
  /queues/{id}/pause:
    post: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/Error' } } }
  /queues/{id}/resume:
    post: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/Error' } } }
  /queues/{id}/close:
    post: { tags: [Queue operations], summary: Close for the day (waiting → cancelled, skipped → no-show), parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK } } }
  /queues/{id}/next:
    post:
      tags: [Queue operations]
      summary: Complete the current patient and call the next one
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses: { '200': { description: '{ completed: "QF-024" | null, called: "QF-025" | null }' }, '409': { description: QUEUE_PAUSED · QUEUE_CLOSED } }
  /queues/{id}/complete:
    post: { tags: [Queue operations], summary: Mark the current patient served, parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/Error' } } }
  /queues/{id}/entries/{entryId}:
    get: { tags: [Queue operations], summary: Patient details + visit timeline (audited), parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }], responses: { '200': { description: OK } } }
  /queues/{id}/entries/{entryId}/call:
    post: { tags: [Queue operations], summary: Call a specific waiting patient, parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }], responses: { '200': { description: OK } } }
  /queues/{id}/entries/{entryId}/skip:
    post: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }], requestBody: { content: { application/json: { schema: { type: object, properties: { reason: { type: string } } } } } }, responses: { '200': { description: OK } } }
  /queues/{id}/entries/{entryId}/requeue:
    post: { tags: [Queue operations], summary: Put a skipped patient back behind the grace positions, parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }], responses: { '200': { description: OK } } }
  /queues/{id}/entries/{entryId}/no-show:
    post: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }], responses: { '200': { description: OK } } }
  /queues/{id}/entries/{entryId}/cancel:
    post: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 3 } } } } } }, responses: { '200': { description: OK } } }
  /queues/{id}/entries/{entryId}/priority:
    post:
      tags: [Queue operations]
      summary: Controlled priority change (reason required, audited)
      parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/entryId' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [priority, reason], properties: { priority: { type: string, enum: [NORMAL, PRIORITY, EMERGENCY] }, reason: { type: string, minLength: 3 } } } } } }
      responses: { '200': { description: OK } }
  /queues/{id}/history:
    get: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }, { name: date, in: query, schema: { type: string, format: date } }, { $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/pageSize' }], responses: { '200': { description: Paged entries } } }
  /queues/{id}/events:
    get: { tags: [Queue operations], parameters: [{ $ref: '#/components/parameters/id' }, { name: limit, in: query, schema: { type: integer, maximum: 200 } }], responses: { '200': { description: Event log, newest first } } }

  # ───────── Patients ─────────
...
  /analytics/queues/{id}:
    get: { tags: [Analytics], parameters: [{ $ref: '#/components/parameters/id' }, { $ref: '#/components/parameters/from' }, { $ref: '#/components/parameters/to' }, { $ref: '#/components/parameters/bucket' }], responses: { '200': { description: OK } } }
...
  /admin/queues:
    get: { tags: [Admin], parameters: [{ name: status, in: query, schema: { type: string, enum: [OPEN, PAUSED, CLOSED] } }], responses: { '200': { description: '{ items: QueueSnapshot[] }' } } }
...